docs: add github_document to literate programming rmd
Now knitr creates a "plain" markdown that renders nicely on github.

The downside is that R code block output are regenerated on every knit.
In order to keep the inline value references in sync with the charts one
would have to not accept certain changes in the md output file if one
does not want to update the png files w/ every git commit.

Furthermore those images are not optimized w/ e.g. pngquant which would cut
the size of PNG file in the current commit size in half.

* add Changelog section
* add `github_document` to `output` as target in YAML so we get a markdown file that renders nicely on Github
* add Output Targets section and sub section on `github_document` output.